fix: remove auto logout, when password is set
Since meteor 1.0.2 the user account is automatically logged out for which the password has been changed.
This behavior is unexpected in the botfront ui and resulted in a bug, where the ui was not able to function because of missing priviliges.
The solution was as simple as setting `logout: false` in the options parameter of the `setPassword` function call.