fix: add empty deps to useTracker to avoid unnecessary model loading
useTracker without dependencies runs every 10seconds reloading and creating a new model object.
As the model is props input for the child components it triggers rerenders. This results in the reset of any unsaved inputs. Upon saving the data within the input field an onchange event is triggered, which results in triggering useTracker. So any updates to the data are still reflected on the ui immediately.