fix: add text property to carousel for new language
When adding a new language for an existing bot, which contains a carousel, the training for the nw language will fail. The generation of the yaml for a language for an existing bot uses a graphql schema to generate it, which lacks the text property. Adding it solves the issue.